Implement client send/receive functions
Also do some gentle refactor of the test/CMakeLists.txt to make it
easier to create test harnesses and one-off tools for validation
behaviors.
